12 - convert wallet

+++ app/src/main/java/com/imdmp/converter/usecase/ConvertCurrencyUseCase.kt
@@ -1,11 +1,9 @@
 package com.imdmp.converter.usecase
 
-import java.math.BigDecimal
-
 interface ConvertCurrencyUseCase {
     suspend operator fun invoke(
         sellData: Double,
         sellCurrencyId: String,
         receiveCurrencyId: String
-    ): BigDecimal
+    ): Double
 }
